requirements.txt
redis>=4.5.0
click>=8.0.0
prometheus-client>=0.16.0
aioredis>=2.0.0
pyyaml>=6.0
psutil>=5.9.0
numpy>=1.24.0
python-dotenv>=1.0.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-exporter-prometheus>=0.55b0
opentelemetry-instrumentation-redis>=0.55b0
opentelemetry-instrumentation>=0.55b0
